refactor(groundcover): make logs/traces explicit @tool functions
Convert GroundcoverLogsTool and GroundcoverTracesTool from the
make_signal_tool factory + __module__ reassignment workaround to explicit
@tool-decorated module functions, matching the SigNoz/QueryReference shape.
Each tool now defines its own _is_available/_extract_params/run in its module
and calls the shared run_signal_query/base_extract_params helpers, so the
registry homes the callables naturally with no __module__ hack.

Delete the now-unused make_signal_tool factory (and its Callable import) from
app/tools/utils/groundcover.py; the genuinely-shared logic helpers stay.

Behavior preserved: same output envelope, groundcover_<signal> source labels,
synthetic-backend short-circuit, credential-free model-facing schema
(additionalProperties:false), required query with seed defaults, and the
corrected gcQL guidance. Add concise docstrings to the public tool functions
and availability/extract helpers for consistency and docstring coverage.
